refactor: alter project structure and remove unnecessary relative imports
