+++ frontend/src/app/PageList.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
-import CreateEvent from './pages/CreateEvent/CreateEvent';
-import Event from './pages/Event/Event';
-import NotFoundPage from './pages/NotFoundPage';
+import CreateEvent from 'pages/CreateEvent/CreateEvent';
+import Event from 'pages/Event/Event';
+import NotFoundPage from 'pages/NotFoundPage';
 
 const PageList = () => {
   return (
