+++ frontend/src/components/ThemeToggle.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { AppContext } from "../contexts";
+import { AppContext } from "global/contexts";
 
 function ThemeToggle() {
   const { theme, setTheme } = React.useContext(AppContext);
