+++ frontend/src/pages/CreateEvent/CreateEvent.tsx
@@ -1,5 +1,5 @@
-import DateSelectionCalendar from "../../components/DateSelectionCalendar/DateSelectionCalendar";
-import Navbar from "../../components/Navbar";
+import DateSelectionCalendar from "./components/DateSelectionCalendar/DateSelectionCalendar";
+import Navbar from "src/components/Navbar";
 import SelectBox from "./components/SelectBox";
 import useCreateEvent from "./hooks/useCreateEvent";
