+++ frontend/src/pages/CreateEvent/components/DateSelectionCalendar/const.ts
