+++ frontend/src/pages/Event/components/AvailabilitiesCalendar/hooks/useUserContext.ts
@@ -1,7 +1,8 @@
 import React from "react";
+import { UserId } from "../../../../../types";
 
 const useUserContext = () => {
-  const [userId, setUserId] = React.useState<string | null>(null);
+  const [userId, setUserId] = React.useState<UserId | null>(null);
   const [userName, setUserName] = React.useState<string>("");
   const [isSaving, setIsSaving] = React.useState(false);
   const [hasConfirmedName, setHasConfirmedName] = React.useState(false);
