+++ frontend/src/pages/NotFoundPage.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import background from "../assets/404.webp";
+import background from "src/assets/404.webp";
 
 const NotFoundPage = () => {
   return (
@@ -21,3 +21,4 @@ const NotFoundPage = () => {
 };
 
 export default NotFoundPage;
+
