--- frontend/src/types/calendar.ts
@@ -1,9 +1,0 @@
-export type Slot = string;
-export type Slots = Set<Slot>;
-export type SlotId = string;
-
-export interface Position {
-  row: number;
-  col: number;
-}
-
