+++ frontend/vite.config.ts
@@ -4,4 +4,11 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      src: "/src",
+      global: "/src/global",
+      pages: "/src/pages",
+    },
+  },
 })
